# packages/api-contract/openapi.yaml
openapi: 3.1.0
info:
  title: PongCode API
  version: 2.0.0
  description: >
    PongCode 前后端分离后的现有 Flask 接口契约。该文件只描述当前接口，
    不改变服务端业务、权限、数据库或文件上传行为。
servers:
  - url: /
...
paths:
...
  /api/projects/{projectId}/bugs:
    parameters:
      - $ref: "#/components/parameters/ProjectId"
    get:
      operationId: listBugs
      parameters:
        - $ref: "#/components/parameters/Search"
        - $ref: "#/components/parameters/Status"
        - name: severity
          in: query
          schema:
            type: integer
        - name: assignee_id
          in: query
          schema:
            oneOf:
              - type: integer
              - type: string
                const: unassigned
      responses:
        "200":
          description: 缺陷列表
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Bug"
    post:
      operationId: createBug
      requestBody: *json-body
      responses: *created-response
  /api/projects/{projectId}/bugs/stats:
    parameters:
      - $ref: "#/components/parameters/ProjectId"
    get:
      operationId: getBugStats
      responses: *json-response
...
  /api/bugs/{bugId}:
    parameters:
      - $ref: "#/components/parameters/BugId"
    get:
      operationId: getBug
      responses: *json-response
    put:
      operationId: updateBug
      requestBody: *json-body
      responses: *json-response
    delete:
      operationId: deleteBug
      responses: *json-response
  /api/bugs/{bugId}/evidences:
    parameters:
      - $ref: "#/components/parameters/BugId"
    post:
      operationId: createBugEvidence
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                comment:
                  type: string
                stack_trace:
                  type: string
                screenshots:
                  type: array
                  maxItems: 5
                  items:
                    type: string
                    format: binary
      responses: *created-response
  /api/bugs/{bugId}/worklogs:
    parameters:
      - $ref: "#/components/parameters/BugId"
    post:
      operationId: createBugWorklog
      requestBody: *json-body
      responses: *created-response
  /api/bugs/{bugId}/worklogs/{worklogId}:
    parameters:
      - $ref: "#/components/parameters/BugId"
      - $ref: "#/components/parameters/WorklogId"
    delete:
      operationId: deleteBugWorklog
      responses: *json-response
...
components:
  parameters:
...
    ProjectId:
      name: projectId
      in: path
      required: true
      schema:
        type: integer
...
    BugId:
      name: bugId
      in: path
      required: true
      schema:
        type: integer
    WorklogId:
      name: worklogId
      in: path
      required: true
      schema:
        type: integer
    Search:
      name: search
      in: query
      schema:
        type: string
    Status:
      name: status
      in: query
      schema:
        type: string
  schemas:
...
    Bug:
      type: object
      required:
        - id
        - title
        - description
        - severity
        - status
        - project_id
      properties:
        id:
          type: integer
        item_code:
          type:
            - string
            - "null"
        title:
          type: string
        description:
          type: string
        severity:
          type: integer
        status:
          type: string
          enum: [open, in_progress, fixed, resolved, closed, rejected]
        project_id:
          type: integer
        sprint_id:
          type:
            - integer
            - "null"
        requirement_id:
          type:
            - integer
            - "null"
        assignee_id:
          type:
            - integer
            - "null"
        evidence_count:
          type: integer
        time_estimate:
          type: number
        time_spent:
          type: number
        bug_type:
          type: string
          enum: [functional, performance, api, security, ui, compatibility, usability, config, data, requirement]
        priority:
          type: string
          enum: [critical, high, normal, low, lowest]
        platform:
          type: string
          enum: [server, h5, android, ios, harmony, pc_web]
        discovery_phase:
          type: string
          enum: [smoke, round_1, round_2, regression, acceptance, integration, gray, production]
        discovery_channel:
          type:
            - string
            - "null"
          enum: [user_feedback, monitoring, log, sprint, null]